Refactored GraphQL Controller Removed Setter in ConfigurationDAO
