+++ src/main/java/gov/nih/nci/bento/model/ConfigurationDAO.java
@@ -4,7 +4,6 @@ import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -53,14 +52,6 @@ public class ConfigurationDAO {
 	//GraphQL
 	@Value("${graphql.schema}")
 	private String schemaFile;
-
-	public String getEsSchemaFile() {
-		return esSchemaFile;
-	}
-
-	public void setEsSchemaFile(String esSchemaFile) {
-		this.esSchemaFile = esSchemaFile;
-	}
 
 	@Value("${graphql.es_schema}")
 	private String esSchemaFile;
